% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fewshot}
\alias{example_fewshot}
\title{Example fewshot examples dataset}
\format{
A data frame with 50 rows and 3 columns:
\describe{
  \item{input}{few shot example test}
  \item{explanation}{explanation for the true label}
  \item{label}{true label}
}
}
\usage{
example_fewshot
}
\description{
A dataset containing few-shot examples for testing functions.
}
\keyword{datasets}
