% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_instructions}
\alias{example_instructions}
\title{Annotation instructions without explanation}
\format{
A character vector.
}
\usage{
example_instructions
}
\description{
Example instructions used for annotation with bacchuss.
}
\keyword{datasets}
