% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_instructions_expl}
\alias{example_instructions_expl}
\title{Annotation instructions for explanation mode}
\format{
A character vector.
}
\usage{
example_instructions_expl
}
\description{
Example instructions used for annotation with bacchuss.
}
\keyword{datasets}
