% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_reminder}
\alias{example_reminder}
\title{Annotation reminder without explanation}
\format{
A character vector.
}
\usage{
example_reminder
}
\description{
Example reminder used for annotation with bacchuss.
}
\keyword{datasets}
