% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_set}
\alias{example_set}
\title{Example fewshot dataset}
\format{
A data frame with 809 rows and 5 columns:
\describe{
  \item{paragraphs}{text for annotation}
  \item{group}{the group mentioned in the text}
  \item{Emotion}{true label from human annotation}
  \item{length}{word count of text}
  \item{estimated context length}{estimated context length for annotation}r
}
}
\usage{
example_set
}
\description{
A dataset containing human annotated data to test annotation functions.
}
\keyword{datasets}
