% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epithets.R
\name{liknites}
\alias{liknites}
\title{Test few-shot examples with leave-one-out annotation}
\usage{
liknites(
  instructions,
  examples,
  codes,
  explanations = NULL,
  n_runs = 3,
  reminder_s = "",
  reminder,
  expected_response_format = c("plain", "json"),
  model,
  host = NULL,
  api_key = NULL,
  backend = c("ollama", "litellm"),
  temperature = 0.2,
  seed = NULL,
  retry_loop = TRUE,
  retry_sleep = 30,
  max_retries = 3,
  default_ctx = 5120,
  verbose = TRUE,
  max_tokens = NULL,
  reasoning = FALSE
)
}
\arguments{
\item{instructions}{A char with your coding instructions}

\item{examples, codes}{Two vectors with annotation examples: the
example text and the correct label for each case}

\item{explanations}{Leave Null if you don't want explanations for
each label.
   * If you use zero shot and want explanations, set TRUE
   * If you use few shot, give a vector with explanations for the
   example labels}

\item{n_runs}{Number of annotation runs over the same examples. Default: 3.}

\item{reminder_s}{Short reminders to send after each coding example.
Can be left empty, does not usually improve reliability and costs
too many tokens}

\item{reminder}{Short reminder sent after all coding examples. Can
Increase stability of labels - repeat instructions for available labels,
rules the model needs reminder of (for example "do not interpret, only
go by what's explicit in the text").}

\item{expected_response_format}{Do you use plaintext or json format? Default plain.}

\item{model}{Which LLM model to use. No default.}

\item{host}{Address of the host you use. Default: local ollama
installation (uses default from ollamar package)}

\item{api_key}{Which API key to use (if your host handes users with
API keys). Default: NULL for unrestricted APIs.}

\item{backend}{Are you using a local Ollama instance or a LiteLLM host?}

\item{temperature}{Which temperature to use for the LLM. Higher =
more creative, lower = more consistent. Anything above 1 is not advised.
Default set to 0.2.}

\item{seed}{Seed to use for deterministic annotation. In unsuccessful
tries, will try again with seed + 1. Default: NULL}

\item{retry_loop}{Small LLMs sometimes get into a long loop producing irrelevant
text. If set TRUE, the results of this run will be discarded and the LLM will
rerun with seed + 1. If set FALSE, keep result from loop - you can inspect the
raw output to check what happened.}

\item{retry_sleep}{How long to wait before retrying - to avoid over-
loading local ollama instances. Default: 30 seconds.}

\item{max_retries}{Max number of retries per call. Usually only needs
one retry. Default: 3.}

\item{default_ctx}{Default context length if no context length is handed
over. Default: 5120.}

\item{verbose}{If update messages are sent. Default: TRUE.}

\item{max_tokens}{Cap length of response - forces the llm to stop generating 
when max token length is reached. Use when llm fails and you suspect infinite
generation of text is the cause.}

\item{reasoning}{If the model allows it, use reasoning. FALSE by default.}
}
\value{
A data frame with added columns:
\describe{

  
  \item{example_id}{An id for each example}
  \item{text}{The example text.}
  \item{true_label}{The original label.}
  \item{labels_runx}{Annotated labels by run number.}
  \item{explanations_runx}{Explanations when requested by run number.}
  \item{majority_label}{Label that got annotated in most runs}
  \item{majority_count}{How often majority label was chosen.}
  \item{agreement}{majority_count / number of runs}
  \item{correct_majority}{Whether the majority call was correct}
}
}
\description{
`liknites()` is a function to test few-shot annotation examples. Using the leave-one-out
method, we test how well the instruction plus all few-shot examples except one would code
the left-out example. 
The output will tell you which examples are informative (by having low agreement) and which
ones are more redundant.
}
